Exercise 4.17
Expand blogs so that each blog contains information on the creator of the blog.

Modify adding new blogs so that when a new blog is created, any user from the database is designated as its creator (for example the one found first). Implement this according to part 4 chapter populate.

Modify listing all blogs so that the creator's user information is displayed with the blog and listing all users also displays the blogs created by each user