Exercises 5.5, 5.6 and 5.7
Change the form for creating blog posts so that it is only displayed when appropriate. It expands when button create new blog is clicked. The form closes when a new blog is created.

Separate the form for creating a new blog into its own component (if you have not already done so), and move all the states required for creating a new blog to this component.

et's add a button to each blog, which controls whether all of the details about the blog are shown or not.

Full details of the blog open when the button is clicked. And the details are hidden when the button is clicked again.

At this point the like button does not need to do anything.