Exercise 6.10
The application has a ready-made body for the Notification component:

Extend the component so that it renders the message stored in the Redux store.

You will have to make changes to the application's existing reducer. Create a separate reducer for the new functionality by using the Redux Toolkit's createSlice function. Also, refactor the application so that it uses a combined reducer as shown in this part of the course material.

The application does not have to use the Notification component in an intelligent way at this point in the exercises. It is enough for the application to display the initial value set for the message in the notificationReducer.